moved utils directory under src/, removed unused code in MosaicSession, renmaed VariantType component to QualitativeBarChart
