+++ src/utils/shortenFloat.js
